requirements.txt
selenium>=4.10.0
undetected-chromedriver>=3.5.0
openai>=1.0.0
playwright>=1.39.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0 